refactor: move arch specific KVM code into modules
Follow up of the refactoring we did in the past around vstate/vm.rs,
where we separate architecture specific code in modules, keeping only
the arch-independent code in 'kvm/mod.rs'.

No functional changes intended.
